+++ tests/test_grid_search.py
@@ -6,6 +6,7 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 import os
+import random
 
 from surprise import Dataset
 from surprise import Reader
@@ -19,10 +20,12 @@ train_file = os.path.join(os.path.dirname(__file__), './u1_ml100k_train')
 test_file = os.path.join(os.path.dirname(__file__), './u1_ml100k_test')
 data = Dataset.load_from_folds([(train_file, test_file)], Reader('ml-100k'))
 
+random.seed(0)
+
 
 def test_grid_search_cv_results():
     param_grid = {'n_epochs': [1, 2], 'lr_all': [0.002, 0.005],
-                  'reg_all': [0.4, 0.6], 'n_factors': [1]}
+                  'reg_all': [0.4, 0.6], 'n_factors': [1], 'init_std_dev': [0]}
     grid_search = GridSearch(SVD, param_grid)
     grid_search.evaluate(data)
     assert len(grid_search.cv_results['params']) == 8
@@ -30,20 +33,22 @@ def test_grid_search_cv_results():
 
 def test_best_rmse_fcp():
     param_grid = {'n_epochs': [1, 2], 'lr_all': [0.002, 0.005],
-                  'reg_all': [0.4, 0.6], 'n_factors': [1]}
+                  'reg_all': [0.4, 0.6], 'n_factors': [1], 'init_std_dev': [0]}
     grid_search = GridSearch(SVD, param_grid, measures=['FCP', 'rmse'])
     grid_search.evaluate(data)
 
     assert grid_search.best_params['RMSE'] == {
-        'lr_all': 0.005, 'n_factors': 1, 'reg_all': 0.4, 'n_epochs': 2}
+        'lr_all': 0.005, 'n_factors': 1, 'reg_all': 0.4, 'n_epochs': 2,
+        'init_std_dev': 0}
 
     assert grid_search.best_params['FCP'] == {
-        'reg_all': 0.6, 'n_epochs': 2, 'lr_all': 0.005, 'n_factors': 1}
+        'reg_all': 0.6, 'n_epochs': 2, 'lr_all': 0.002, 'n_factors': 1,
+        'init_std_dev': 0}
 
 
 def test_measure_is_not_case_sensitive():
     param_grid = {'n_epochs': [1], 'lr_all': [0.002, 0.005],
-                  'reg_all': [0.4, 0.6], 'n_factors': [1]}
+                  'reg_all': [0.4, 0.6], 'n_factors': [1], 'init_std_dev': [0]}
     grid_search = GridSearch(SVD, param_grid, measures=['FCP', 'mae', 'rMSE'])
     grid_search.evaluate(data)
     assert grid_search.best_index['fcp'] == grid_search.best_index['FCP']
@@ -53,7 +58,7 @@ def test_measure_is_not_case_sensitive():
 
 def test_best_estimator():
     param_grid = {'n_epochs': [5], 'lr_all': [0.002, 0.005],
-                  'reg_all': [0.4, 0.6], 'n_factors': [1]}
+                  'reg_all': [0.4, 0.6], 'n_factors': [1], 'init_std_dev': [0]}
     grid_search = GridSearch(SVD, param_grid, measures=['FCP', 'mae', 'rMSE'])
     grid_search.evaluate(data)
     best_estimator = grid_search.best_estimator['MAE']